chore(file-manager): move access token logic to common-server
