+++ secret-manager/api/api.go
@@ -8,12 +8,12 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/telekom/controlplane/common-server/api/accesstoken"
 	"net/http"
 	"os"
 	"strings"
 
 	"github.com/pkg/errors"
-	"github.com/telekom/controlplane/secret-manager/api/accesstoken"
 	"github.com/telekom/controlplane/secret-manager/api/gen"
 	"github.com/telekom/controlplane/secret-manager/api/util"
 )
