+++ tools/snapshotter/main.go
@@ -16,11 +16,11 @@ import (
 	"github.com/goccy/go-yaml"
 	"github.com/joho/godotenv"
 	"github.com/pkg/errors"
+	"github.com/telekom/controlplane/common-server/api/accesstoken"
 	gatewayv1 "github.com/telekom/controlplane/gateway/api/v1"
 	kong "github.com/telekom/controlplane/gateway/pkg/kong/api"
 	"github.com/telekom/controlplane/gateway/pkg/kongutil"
 	secrets "github.com/telekom/controlplane/secret-manager/api"
-	"github.com/telekom/controlplane/secret-manager/api/accesstoken"
 	"github.com/telekom/controlplane/tools/snapshotter/getters"
 	"github.com/telekom/controlplane/tools/snapshotter/state"
 	"github.com/telekom/controlplane/tools/snapshotter/util"
